backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
tensorflow-cpu==2.19.0
numpy==1.26.4
pydantic==2.6.4
joblib==1.5.2
scikit-learn==1.6.1